// src/app/models/client-model/client.ts
export class Client {

  name: string = '';
  lastname: string = '';
  cpf: string = '';
  cellphone: string = '';
  email: string = '';
  password: string = '';
  password2: string = '';
  cep: string = '';
  street_name: string = '';
  state: string = '';
  city: string = '';
  additional_info: string = '';
  
}